requirements.txt
PyYAML~=6.0
numpy==1.22.4
bitarray~=2.5.1